+++ CHANGES.rst
@@ -11,6 +11,10 @@ in progress
 - Added support for Ubuntu 24 "Noble Numbat"
 - Updated OS registry with Fedora 44 and 45, CentOS 10, RHEL 10, RockyLinux 9,
   OpenSUSE Leap 16, AmazonLinux 2023, Oracle Linux 9 and 10
+- Make Windows runner subsystem production ready.
+- Add support for Windows Server Core 2019 and friends, like
+  ``windows/servercore:ltsc2019``, ``windows/nanoserver:1809``, or
+  ``eclipse-temurin:17-jdk``.
 
 2023-02-14 0.3.0
 ================
@@ -21,7 +25,6 @@ in progress
 - Improve central command invocation function
 - Improve documentation
 - Fix packaging
-
 
 2022-05-20 0.2.0
 ================